Split import script to 2 files for easier maintenance
